requirements.txt
scikit_learn==0.21.0
flask==1.0.2
Flask-RESTful==0.3.7
keras==2.2.4
tensorflow==1.13.1
google-cloud-storage==1.16.0